Renamed drawing to graphics and extracted display-specific stuff into the display folders
