+++ src/epd4in2/graphics.rs
@@ -1,39 +1,21 @@
+use graphics::{
+    outside_display,
+    rotation,
+    DisplayRotation, 
+    Display
+};
 use color::Color;
 use embedded_graphics::prelude::*;
 
-/// Displayrotation 
-#[derive(Clone, Copy)]
-pub enum DisplayRotation {
-    /// No rotation
-    Rotate0,
-    /// Rotate by 90 degrees clockwise
-    Rotate90,
-    /// Rotate by 180 degrees clockwise
-    Rotate180,
-    /// Rotate 270 degrees clockwise
-    Rotate270,
-}
-impl Default for DisplayRotation {
-    fn default() -> Self {
-        DisplayRotation::Rotate0
-    }
-}
+use epd4in2::constants::{DEFAULT_BACKGROUND_COLOR, WIDTH, HEIGHT};
 
-pub trait Display {
-    fn buffer(&self) -> &[u8];
-    fn set_rotation(&mut self, rotation: DisplayRotation);
-    fn rotation(&self) -> DisplayRotation;
-}
-
-
-pub struct DisplayEink42BlackWhite {
-    buffer: [u8; 400 * 300 / 8],
+pub struct DisplayEink42BlackWhite {    
+    buffer: [u8; WIDTH as usize * HEIGHT as usize / 8],
     rotation: DisplayRotation,
 }
 
 impl Default for DisplayEink42BlackWhite {
     fn default() -> Self {
-        use epd4in2::constants::{DEFAULT_BACKGROUND_COLOR, WIDTH, HEIGHT};
         DisplayEink42BlackWhite {
             buffer: [
                 DEFAULT_BACKGROUND_COLOR.get_byte_value();
@@ -62,8 +44,6 @@ impl Drawing<Color> for DisplayEink42BlackWhite {
     where
         T: Iterator<Item = Pixel<Color>>
     {
-        use epd4in2::constants::{WIDTH, HEIGHT};
-
         let width = WIDTH as u32;
         let height = HEIGHT as u32;
 
@@ -87,45 +67,6 @@ impl Drawing<Color> for DisplayEink42BlackWhite {
     }
 }
 
-fn outside_display(x: u32, y: u32, width: u32, height: u32, rotation: DisplayRotation) -> bool {
-    match rotation {
-        DisplayRotation::Rotate0 | DisplayRotation::Rotate180 => {
-            if x >= width || y >= height {
-                return true;
-            }
-        },
-        DisplayRotation::Rotate90 | DisplayRotation::Rotate270 => {
-            if y >= width || x >= height {
-                return true;
-            } 
-        }
-    }
-    false
-}
-
-fn rotation(x: u32, y: u32, width: u32, height: u32, rotation: DisplayRotation) -> (u32, u8) {
-    match rotation {
-        DisplayRotation::Rotate0 => (
-            x / 8 + (width / 8) * y,
-            0x80 >> (x % 8),
-        ),
-        DisplayRotation::Rotate90 => (
-            (width - 1 - y) / 8 + (width / 8) * x,
-            0x01 << (y % 8),
-        ),
-        DisplayRotation::Rotate180 => (
-            ((width / 8) * height - 1) - (x / 8 + (width / 8) * y),
-            0x01 << (x % 8),
-        ),
-        DisplayRotation::Rotate270 => (
-            y / 8 + (height - 1 - x) * (width / 8),
-            0x80 >> (y % 8),
-        ),
-    }
-}
-
-
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -133,42 +74,16 @@ mod tests {
     use embedded_graphics::coord::Coord;
     use embedded_graphics::primitives::Line;
 
-    #[test]
-    fn rotation_overflow() {
-        use epd4in2::constants::{WIDTH, HEIGHT};
-        let width = WIDTH as u32;
-        let height = HEIGHT as u32;
-        test_rotation_overflow(width, height, DisplayRotation::Rotate0);
-        test_rotation_overflow(width, height, DisplayRotation::Rotate90);
-        test_rotation_overflow(width, height, DisplayRotation::Rotate180);
-        test_rotation_overflow(width, height, DisplayRotation::Rotate270);
-        
-    }
-
-    fn test_rotation_overflow(width: u32, height: u32, rotation2: DisplayRotation) {
-        let max_value = width / 8 * height;
-        for x in 0..(width + height) { //limit x because it runs too long 
-            for y in 0..(u32::max_value()) {
-                if outside_display(x, y, width, height, rotation2) {
-                    break;
-                } else {
-                    let (idx, _) = rotation(x, y, width, height, rotation2);
-                    assert!(idx < max_value);
-                }
-            }
-        }
-    }
-
     // test buffer length
     #[test]
-    fn graphics_4in2_size() {
+    fn graphics_size() {
         let display = DisplayEink42BlackWhite::default();
         assert_eq!(display.buffer().len(), 15000);
     }
     
     // test default background color on all bytes
     #[test]
-    fn graphics_4in2_default() {
+    fn graphics_default() {
         let display = DisplayEink42BlackWhite::default();
         use epd4in2;
         for &byte in display.buffer() {
@@ -177,7 +92,7 @@ mod tests {
     }
 
     #[test]
-    fn graphics_4in2_rotation_0() {
+    fn graphics_rotation_0() {
         let mut display = DisplayEink42BlackWhite::default();
         display.draw(
             Line::new(Coord::new(0, 0), Coord::new(7, 0))
@@ -195,7 +110,7 @@ mod tests {
     }
 
     #[test]
-    fn graphics_4in2_rotation_90() {
+    fn graphics_rotation_90() {
         let mut display = DisplayEink42BlackWhite::default();
         display.set_rotation(DisplayRotation::Rotate90);
         display.draw(
@@ -214,7 +129,7 @@ mod tests {
     }
 
     #[test]
-    fn graphics_4in2_rotation_180() {
+    fn graphics_rotation_180() {
         let mut display = DisplayEink42BlackWhite::default();
         display.set_rotation(DisplayRotation::Rotate180);
         display.draw(
@@ -237,7 +152,7 @@ mod tests {
     }
 
     #[test]
-    fn graphics_4in2_rotation_270() {
+    fn graphics_rotation_270() {
                 let mut display = DisplayEink42BlackWhite::default();
         display.set_rotation(DisplayRotation::Rotate270);
         display.draw(
